+++ compiler/src/org.graalvm.compiler.nodes/src/org/graalvm/compiler/nodes/InvokeNode.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2009, 2016, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2009, 2018, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -67,9 +67,9 @@ import jdk.vm.ci.meta.JavaKind;
           allowedUsageTypes = {Memory},
           cycles = CYCLES_UNKNOWN,
           cyclesRationale = "We cannot estimate the runtime cost of a call, it is a blackhole." +
-                            "However, we can estimate, dyanmically, the cost of the call operation itself based on the type of the call.",
+                            "However, we can estimate, dynamically, the cost of the call operation itself based on the type of the call.",
           size = SIZE_UNKNOWN,
-          sizeRationale = "We can only dyanmically, based on the type of the call (special, static, virtual, interface) decide" +
+          sizeRationale = "We can only dynamically, based on the type of the call (special, static, virtual, interface) decide" +
                           "how much code is generated for the call.")
 // @formatter:on
 public final class InvokeNode extends AbstractMemoryCheckpoint implements Invoke, LIRLowerable, MemoryCheckpoint.Single, UncheckedInterfaceProvider {
@@ -81,17 +81,27 @@ public final class InvokeNode extends AbstractMemoryCheckpoint implements Invoke
     protected final int bci;
     protected boolean polymorphic;
     protected boolean useForInlining;
+    protected LocationIdentity identity;
 
     public InvokeNode(CallTargetNode callTarget, int bci) {
         this(callTarget, bci, callTarget.returnStamp().getTrustedStamp());
     }
 
+    public InvokeNode(CallTargetNode callTarget, int bci, LocationIdentity identity) {
+        this(callTarget, bci, callTarget.returnStamp().getTrustedStamp(), identity);
+    }
+
     public InvokeNode(CallTargetNode callTarget, int bci, Stamp stamp) {
+        this(callTarget, bci, stamp, LocationIdentity.any());
+    }
+
+    public InvokeNode(CallTargetNode callTarget, int bci, Stamp stamp, LocationIdentity identity) {
         super(TYPE, stamp);
         this.callTarget = callTarget;
         this.bci = bci;
         this.polymorphic = false;
         this.useForInlining = true;
+        this.identity = identity;
     }
 
     @Override
@@ -158,7 +168,7 @@ public final class InvokeNode extends AbstractMemoryCheckpoint implements Invoke
 
     @Override
     public LocationIdentity getLocationIdentity() {
-        return LocationIdentity.any();
+        return identity;
     }
 
     @Override