+++ substratevm/mx.substratevm/mx_substratevm.py
@@ -382,7 +382,8 @@ def layout_native_image_root(native_image_root):
                 symlink_or_copy(join(jvmci_path, symlink_name), join(native_image_root, 'lib', 'jvmci', symlink_name))
 
     # Create native-image layout for truffle parts
-    native_image_layout_dists(join('lib', 'truffle'), ['truffle:TRUFFLE_API', 'truffle:TRUFFLE_NFI'])
+    if mx.get_os() != 'windows':  # necessary until Truffle is fully supported (GR-7941)
+        native_image_layout_dists(join('lib', 'truffle'), ['truffle:TRUFFLE_API', 'truffle:TRUFFLE_NFI'])
 
     # Create native-image layout for tools parts
     for tool_name in tools_map: