+++ sulong/projects/com.oracle.truffle.llvm.nodes/src/com/oracle/truffle/llvm/nodes/op/ToComparableValue.java
@@ -65,7 +65,7 @@ public abstract class ToComparableValue extends LLVMNode {
         return NativeToComparableValueNodeGen.create();
     }
 
-    @TruffleBoundary
+    @TruffleBoundary(allowInlining = true)
     private static long getHashCode(Object obj) {
         // if we ever switch to a more robust implementation, we can simplify the
         // LLVMManagedCompareNode